Refactor into separate types module
Also use `fmt::Display` instead of ToString where possible
